% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_httpuv.R
\name{demo_httpuv}
\alias{demo_httpuv}
\title{Demo multipart parser with httpuv}
\usage{
demo_httpuv(port = 9359)
}
\arguments{
\item{port}{which port number to run the http server}
}
\description{
Starts the httpuv web server and hosts a simple form including a file
upload to demo the multipart parser.
}
\seealso{
Other demo:
\code{\link[=demo_rhttpd]{demo_rhttpd()}}
}
\concept{demo}
