% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_boundary}
\alias{get_boundary}
\title{Extract boundary from header}
\usage{
get_boundary(content_type)
}
\arguments{
\item{content_type}{the Content-Type request header}
}
\description{
Helper function to extract the boundary value from the content-type
http request header.
}
\examples{
content_type = "multipart/form-data; boundary=----WebKitFormBoundaryxEQWPBrp9Wym1xpB"
get_boundary(content_type)
}

