% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_query}
\alias{parse_query}
\title{Parse query string}
\usage{
parse_query(query)
}
\description{
Parse a query string (the part after the question mark of a URL). This
includes decoding of URL encoded values.
}
\examples{
myquery <- "foo=1\%2B1\%3D2&bar=yin\%26yang"
parse_query(myquery)
}

