% Generated by roxygen2 (4.0.2): do not edit by hand
\name{demo_httpuv}
\alias{demo_httpuv}
\title{Demo multipart parser with httpuv}
\usage{
demo_httpuv(port)
}
\arguments{
\item{port}{which port number to run the http server}
}
\description{
Starts the httpuv web server and hosts a simple form including a file
upload to demo the multipart parser.
}

