% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_rhttpd.R
\name{demo_rhttpd}
\alias{demo_rhttpd}
\title{Demo multipart parser with rhttpd}
\usage{
demo_rhttpd()
}
\description{
Starts the Rhttpd web server and hosts a simple form including a file
upload to demo the multipart parser.
}
\seealso{
Other demo: 
\code{\link{demo_httpuv}()}
}
\concept{demo}
