% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_multipart}
\alias{parse_multipart}
\title{Parse a multipart/form-data request}
\usage{
parse_multipart(body, boundary)
}
\arguments{
\item{body}{body of the HTTP request. Must be raw or character vector.}

\item{boundary}{boundary string as specified in the \code{Content-Type} request header.}
}
\description{
This function parses a multipart/form-data request, which is usually generated as
part of an HTML form submission. The result can contain both text values as well
as (binary) files. They can be distinguished from the \code{filename} attribute.
}
\details{
A multipart/form-data request consists of a single body which contains one or more
values separated with a boundary string. This boundary string is chosen by the client
(e.g. the browser) and specified in the \code{Content-Type} header of the HTTP request.

The parse is written in pure R, but still pretty fast.
}
\examples{
\dontrun{
# Launch the example page
demo_rhttpd()
}
}

